Add files via upload
The patcher in its state as of the demonstration video on Jan 27 2022: https://twitter.com/bootyjazz/status/1486813442640121862

Requires Max 8 to run. There are a couple important issues with autocutting logic that need to be fixed. However, using it manually works, and it can spit out a batch file for you.